+++ src/js/App/GlobalFilter/constants.js
@@ -1,11 +1,8 @@
-import { deleteLocalStorageItems } from '../../utils';
 import { decodeToken, GLOBAL_FILTER_KEY } from '../../jwt/jwt';
 import omit from 'lodash/omit';
 import flatMap from 'lodash/flatMap';
 import memoize from 'lodash/memoize';
 import { SID_KEY, AAP_KEY, MSSQL_KEY } from '../../redux/globalFilterReducers';
-import { spinUpStore } from '../../redux-config';
-import { storeInitialHash } from '../../redux/actions';
 
 export const INVENTORY_API_BASE = '/api/inventory/v1';
 export const workloads = [
@@ -58,60 +55,6 @@ export const createTagsFilter = (tags = []) =>
       },
     };
   }, {});
-
-export const storeFilter = (tags, token, isEnabled, history, firstLoad) => {
-  deleteLocalStorageItems(Object.keys(localStorage).filter((key) => key.startsWith(GLOBAL_FILTER_KEY)));
-  if (isEnabled) {
-    const searchParams = new URLSearchParams();
-    const [, SIDs, mappedTags] = flatTags(tags, false, true);
-    if (tags?.Workloads) {
-      const currWorkloads = Object.entries(tags?.Workloads || {})?.find(([, { isSelected }]) => isSelected)?.[0];
-      if (currWorkloads) {
-        searchParams.append('workloads', currWorkloads);
-      }
-    }
-    searchParams.append('SIDs', SIDs);
-    searchParams.append('tags', mappedTags);
-
-    if (firstLoad && window.location.hash.length > 0) {
-      const { store } = spinUpStore();
-      store.dispatch(storeInitialHash(window.location.hash));
-    }
-    history.push({
-      ...history.location,
-      hash: searchParams.toString(),
-    });
-  }
-
-  localStorage.setItem(
-    `${GLOBAL_FILTER_KEY}/${token}`,
-    JSON.stringify(
-      Object.entries(tags).reduce(
-        (acc, [key, value]) => ({
-          ...acc,
-          [key]: {
-            ...Object.entries(value || {}).reduce(
-              (
-                currValue,
-                // eslint-disable-next-line no-unused-vars
-                [itemKey, { item, value: tagValue, group: { items, ...group } = {}, ...rest }]
-              ) => ({
-                ...currValue,
-                [itemKey]: {
-                  ...rest,
-                  item: { tagValue: item?.tagValue || tagValue, tagKey: item?.tagKey || itemKey },
-                  group,
-                },
-              }),
-              {}
-            ),
-          },
-        }),
-        {}
-      )
-    )
-  );
-};
 
 export const generateFilter = async () => {
   const searchParams = new URLSearchParams(location.hash?.substring(1));
