+++ src/js/App/Sidenav/Navigation/Loader.tsx
